move command runner interface outside _private
