+++ python/ray/autoscaler/command_runner.py
@@ -1,0 +1,155 @@
+from shlex import quote
+from typing import Any, List, Tuple, Dict, Optional
+import json
+import os
+import sys
+
+# How long to wait for a node to start, in seconds
+NODE_START_WAIT_S = 300
+HASH_MAX_LENGTH = 10
+KUBECTL_RSYNC = os.path.join(
+    os.path.dirname(os.path.abspath(__file__)), "kubernetes/kubectl-rsync.sh")
+
+_config = {"use_login_shells": True, "silent_rsync": True}
+
+
+def is_rsync_silent():
+    return _config["silent_rsync"]
+
+
+def set_rsync_silent(val):
+    """Choose whether to silence rsync output.
+
+    Most commands will want to list rsync'd files themselves rather than
+    print the default rsync spew.
+    """
+    _config["silent_rsync"] = val
+
+
+def is_using_login_shells():
+    return _config["use_login_shells"]
+
+
+def set_using_login_shells(val):
+    """Choose between login and non-interactive shells.
+
+    Non-interactive shells have the benefit of receiving less output from
+    subcommands (since progress bars and TTY control codes are not printed).
+    Sometimes this can be significant since e.g. `pip install` prints
+    hundreds of progress bar lines when downloading.
+
+    Login shells have the benefit of working very close to how a proper bash
+    session does, regarding how scripts execute and how the environment is
+    setup. This is also how all commands were ran in the past. The only reason
+    to use login shells over non-interactive shells is if you need some weird
+    and non-robust tool to work.
+
+    Args:
+        val (bool): If true, login shells will be used to run all commands.
+    """
+    _config["use_login_shells"] = val
+
+
+def _with_environment_variables(cmd: str,
+                                environment_variables: Dict[str, object]):
+    """Prepend environment variables to a shell command.
+
+    Args:
+        cmd (str): The base command.
+        environment_variables (Dict[str, object]): The set of environment
+            variables. If an environment variable value is a dict, it will
+            automatically be converted to a one line yaml string.
+    """
+
+    as_strings = []
+    for key, val in environment_variables.items():
+        val = json.dumps(val, separators=(",", ":"))
+        s = "export {}={};".format(key, quote(val))
+        as_strings.append(s)
+    all_vars = "".join(as_strings)
+    return all_vars + cmd
+
+
+def _with_interactive(cmd):
+    force_interactive = (
+        f"true && source ~/.bashrc && "
+        f"export OMP_NUM_THREADS=1 PYTHONWARNINGS=ignore && ({cmd})")
+    return ["bash", "--login", "-c", "-i", quote(force_interactive)]
+
+
+class CommandRunnerInterface:
+    """Interface to run commands on a remote cluster node.
+
+    Command runner instances are returned by provider.get_command_runner()."""
+
+    def run(
+            self,
+            cmd: str = None,
+            timeout: int = 120,
+            exit_on_fail: bool = False,
+            port_forward: List[Tuple[int, int]] = None,
+            with_output: bool = False,
+            environment_variables: Dict[str, object] = None,
+            run_env: str = "auto",
+            ssh_options_override_ssh_key: str = "",
+            shutdown_after_run: bool = False,
+    ) -> str:
+        """Run the given command on the cluster node and optionally get output.
+
+        WARNING: the cloudgateway needs arguments of "run" function to be json
+            dumpable to send them over HTTP requests.
+
+        Args:
+            cmd (str): The command to run.
+            timeout (int): The command timeout in seconds.
+            exit_on_fail (bool): Whether to sys exit on failure.
+            port_forward (list): List of (local, remote) ports to forward, or
+                a single tuple.
+            with_output (bool): Whether to return output.
+            environment_variables (Dict[str, str | int | Dict[str, str]):
+                Environment variables that `cmd` should be run with.
+            run_env (str): Options: docker/host/auto. Used in
+                DockerCommandRunner to determine the run environment.
+            ssh_options_override_ssh_key (str): if provided, overwrites
+                SSHOptions class with SSHOptions(ssh_options_override_ssh_key).
+            shutdown_after_run (bool): if provided, shutdowns down the machine
+            after executing the command with `sudo shutdown -h now`.
+        """
+        raise NotImplementedError
+
+    def run_rsync_up(self,
+                     source: str,
+                     target: str,
+                     options: Optional[Dict[str, Any]] = None) -> None:
+        """Rsync files up to the cluster node.
+
+        Args:
+            source (str): The (local) source directory or file.
+            target (str): The (remote) destination path.
+        """
+        raise NotImplementedError
+
+    def run_rsync_down(self,
+                       source: str,
+                       target: str,
+                       options: Optional[Dict[str, Any]] = None) -> None:
+        """Rsync files down from the cluster node.
+
+        Args:
+            source (str): The (remote) source directory or file.
+            target (str): The (local) destination path.
+        """
+        raise NotImplementedError
+
+    def remote_shell_command_str(self) -> str:
+        """Return the command the user can use to open a shell."""
+        raise NotImplementedError
+
+    def run_init(self, *, as_head: bool, file_mounts: Dict[str, str]) -> None:
+        """Used to run extra initialization commands.
+
+        Args:
+            as_head (bool): Run as head image or worker.
+            file_mounts (dict): Files to copy to the head and worker nodes.
+        """
+        pass
